# src/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoonBox API",
    "description": "MoonBox REST API service.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chat/governance-preparations": {
      "post": {
        "tags": [
          "Governance"
        ],
        "summary": "准备治理动作与个人会话",
        "operationId": "create_preparation_api_v1_chat_governance_preparations_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreparationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/conversations/{cid}/governance-candidates": {
      "get": {
        "tags": [
          "Governance"
        ],
        "summary": "读取不可变治理成果",
        "operationId": "get_candidates_api_v1_chat_conversations__cid__governance_candidates_get",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cid"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/governance-candidates/{candidate_id}/applications": {
      "post": {
        "tags": [
          "Governance"
        ],
        "summary": "申请应用固定治理成果",
        "operationId": "post_application_api_v1_chat_governance_candidates__candidate_id__applications_post",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/governance-applications/{operation_id}": {
      "get": {
        "tags": [
          "Governance"
        ],
        "summary": "读取应用与恢复状态",
        "operationId": "get_application_api_v1_chat_governance_applications__operation_id__get",
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Operation Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiResponse_dict_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse[dict]"
      },
...
      "ApplicationRequest": {
        "properties": {
          "expected_manifest_hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Expected Manifest Hash"
          },
          "candidate_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Candidate Revision"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Idempotency Key"
          },
          "maintenance_confirmed": {
            "type": "boolean",
            "title": "Maintenance Confirmed"
          }
        },
        "type": "object",
        "required": [
          "expected_manifest_hash",
          "candidate_revision",
          "idempotency_key",
          "maintenance_confirmed"
        ],
        "title": "ApplicationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreparationRequest": {
        "properties": {
          "space_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Space Id"
          },
          "repository_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Repository Id"
          },
          "object_id": {
            "type": "string",
            "pattern": "^REQ-\\d{4}(?:-[a-z0-9-]+)?$",
            "title": "Object Id"
          },
          "action": {
            "type": "string",
            "const": "req-generate",
            "title": "Action",
            "default": "req-generate"
          }
        },
        "type": "object",
        "required": [
          "space_id",
          "repository_id",
          "object_id"
        ],
        "title": "PreparationRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}